change Integer to Long in all repositories

+++ final_project_spring/src/main/java/extclasses/final_project_spring/repository/ShelfRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ShelfRepository extends JpaRepository<Shelf, Integer> {
+public interface ShelfRepository extends JpaRepository<Shelf, Long> {
     Optional<Shelf> findByBookIsNull();
 }
